requirements.txt
aiogram==3.0.0b6
SQLAlchemy==2.0.0
psycopg2-binary==2.9.6   # если планируете PostgreSQL
python-dotenv==1.0.0
requests==2.28.2
